Extract shared ArkVisionRestClient to deduplicate the settings interfaces
The MJPEG and RTSP ArkVision settings interfaces duplicated the REST transport,
the sensor-mode/framerate tables and the fps mapping. Move all of it into a new
rosys/vision/arkvision_rest_client.py and have both interfaces subclass
ArkVisionRestClient, keeping only their stream-specific bits (resolution/mirror
for MJPEG, bitrate for RTSP). Behaviour is unchanged.
